Update setup_pulse_node.sh
Removed the option to choose between a single script or separate scripts as it has become deprecated. This is because the Docker containers are now set up to restart automatically after the initial startup. Instead, we will generate two separate scripts: start_execution.sh and start_consensus.sh. These scripts will be placed in the directory '$CUSTOM_PATH' and will start the execution and consensus clients respectivel